refactor: isFindSameStation 변수 제거 및 addSection 처리 순서 변경
