+++ docs/README.md
@@ -78,6 +78,6 @@
 * [ ] 인수테스트 코드의 재사용성을 위한 리팩터링
 * [ ] distance 객체로 래핑
 * [ ] 인수테스트 검증시 노선을 조회하도록 수정
-* [ ] isFindSameStation 변수 제거
-* [ ] 종점에 대한 판단로직 다시 작성
+* [x] isFindSameStation 변수 제거
+* [x] 종점에 대한 판단로직 다시 작성
 * [ ] sortNo 의 사용 대신 정렬순서 해결해보기
